remove item_text.c deal with the consequences later
